backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi
email-validator